Avoid stack walk as specified in the exclusion list
